% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{akcan}
\alias{akcan}
\title{Alaska and western Canada.}
\format{A SpatialPolygonsDataFrame with a single feature.}
\usage{
akcan
}
\description{
A vector map of full SNAP PRISM domain, a single union polygon of the State of Alaska and the five SNAP-relevant western Canadian provinces: Alberta, British Columbia, Manitoba, Saskatchewan and Yukon Territory.
}
\keyword{datasets}
