% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{akcan2}
\alias{akcan2}
\title{Alaska and western Canadian provinces.}
\format{A SpatialPolygonsDataFrame with 6 features. See \code{akcan2$NAME}.}
\usage{
akcan2
}
\description{
A vector map like \code{\link{akcan}}, but retaining the six distinct encompassing state/province polygons.
}
\keyword{datasets}
