% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoreg1}
\alias{ecoreg1}
\title{Alaska level 1 ecological regions (ecoregions).}
\format{A SpatialPolygonsDataFrame with 3 features. See \code{ecoreg1$NAME}.}
\usage{
ecoreg1
}
\description{
A vector map of Alaska level 1 ecoregions.
}
\keyword{datasets}
