% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoreg2}
\alias{ecoreg2}
\title{Alaska level 2 ecological regions (ecoregions).}
\format{A SpatialPolygonsDataFrame with 9 features. See \code{ecoreg2$NAME}.}
\usage{
ecoreg2
}
\description{
A vector map of Alaska level 2 ecoregions.
}
\keyword{datasets}
