% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoreg3}
\alias{ecoreg3}
\title{Alaska level 3 ecological regions (ecoregions).}
\format{A SpatialPolygonsDataFrame with 32 features. See \code{ecoreg3$NAME}.}
\usage{
ecoreg3
}
\description{
A vector map of Alaska level 3 ecoregions.
}
\keyword{datasets}
