% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mapdata}
\alias{mapdata}
\title{Basic metadata for all data sets in maputils}
\usage{
mapdata()
}
\value{
a data frame.
}
\description{
This function returns a data frame with basic meta data for all data sets in \code{maputils}.
This includes data object names, spatial domain, number of features, number of variables, ID column name,
whether or not the SpatialPolygonsDataFrame has multilevel/overlapping features, and a short description.
}
\examples{
mapdata()
}
